tdf#164904 RTF import: fix unexpected CharFontFamily
Open the RTF in Word, the font fallback is serif, open in Writer, our
fallback is sans.

\froman is not parsed by the RTF import, but the default font family was
ROMAN and that worked in the past.

Fix the problem by only setting the property in case the tokenizer has a
value for it, there is a difference between not setting it vs setting it
to "don't know".

Regression from commit d06de2e049761b7b9e8a95f17557d309812f7acc
(Related: tdf#162072 DOCX import: handle font family for characters,
2024-07-19).

Change-Id: I4253ac7e27fab7d3e8e61c43659589b4fc177d6c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180934
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>