tdf#159377 sw: fix undo of paste table into footnote
Tables aren't allowed in footnotes, hence pasting a table into a
footnote will paste only the text nodes in the cells but not the table
nodes.

This code isn't prepared for this situation, so the adjustment of the
position doesn't happen.

        if (pCopyPam->GetPoint()->GetNode().IsStartNode())
            pCopyPam->GetPoint()->Adjust(SwNodeOffset(-1));

But then the corresponding adjustment in SwUndoInserts::SetInsertRange()
does happen and the stored start node index is wrong.

It looks like both of these adjustments can simply be removed.

(regression from commit fcd4222d36e1864452163e5c94976eea353bbaf0)

Change-Id: Ib419b29da552cf1df7150178924c45e3743cd7d9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180947
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins