cool#11072: Send current cell's decimal separator.
Calc: In a file with a language that uses "," as the decimal separator,
using a keyboard with a layout that uses "." as decimal separator is difficult.
Numbers are considered as string when it happens.
Users should be able to use NumPad decimal separator regardless of the cell's current decimal separator.

Change-Id: I35623661acb0f505597458b7ce55cfc2c3a4dace
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181993
Reviewed-by: Gökay ŞATIR <gokaysatir@collabora.com>
Tested-by: Jenkins