browsebox a11y: Use comphelper::OAccessibleComponentHelper
Let AccessibleBrowseBoxBase derive from
comphelper::OAccessibleComponentHelper instead
of duplicating much of the logic for notifying of
a11y events and location/bounds handling.

The involved classes are the a11y classes for both,
BrowseBox (see BrowseBox::CreateAccessible)
and SvHeaderTabListBox (see SvHeaderTabListBox::CreateAccessible).

With the fixes to the involved classes from previous
commits in place, a quick test for both of these with
Accerciser and the qt6 VCL plugin didn't show any change in
behavior with this commit in place.

Sample scenario for SvHeaderTabListBox:
"Tools" -> "Options" -> "Advanced" -> "Open Expert Configuration"

Sample scenario for BrowseBox:

1) Start Base, create a new database
2) With the "Tables" section enabled, click on
   "Create Table in Design View..."
3) in Accerciser's treeview of the LO a11y, navigate
   through the a11y objects of the table and its children

A quick test with Orca also didn't reveal any changes
in behavior for the expert configuration case, but triggered
a crash for the BrowseBox case.

The underlying problem is a preexisting issue however, and
will be addressed in an upcoming commit.

Change-Id: I2a540ade0c5f6a5a82c764de4cef37be556453fd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181153
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>