tdf#158198 sw: prevent xBookmark.getAnchor().setString("") from deleting
... the bookmark.

Add a RANGE_IS_BOOKMARK to SwXTextRange and special case setString() so
that it will do a replace, not a delete, to prevent the bookmark from
being deleted.

SwXBookmark::getAnchor() will create such a SwXTextRange.

Bookmarks may be in table cells of course so we need RANGE_IN_CELL to be
a flag now so it can be used with 2 different enum values.

(regression from commit baf8d2c1c16cb3bdc4edad2560f95fea807a034f)

Change-Id: I69169b9a5f6f7525715122b16de009c894246bef
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180985
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins