vcl: Merge LclHeaderTabListBox into SvHeaderTabListBox
By now LclHeaderTabListBox is the only SvHeaderTabListBox
subclass and SvHeaderTabListBox is not used by itself.
Merge the former into the latter to make the class hierarchy
a little less complex.

The "plain" SvHeaderTabListBox or other subclasses used
to be used in the past, but that is no longer the case,
see e.g.

    commit 4e07b178f58b1dabbb48bc9f26155b8270a4a72d
    Date:   Tue Feb 18 10:16:40 2020 +0000

        weld WatchWindow panel

which eliminated one use.

Copy the comment about the default NotifyStartDrag
implementation in the base class being weird from
vcl/inc/treeglue.hxx but also leave it there, as
it applies for the LclTabListBox class that is
still there just the same.

Change-Id: I658332a9675a9c67c7250f2577a6eb1a0cc491f7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181022
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>