tdf#164996 Optimize Persian (Farsi) AutoCorrect list
This patch provides an optimized version of the AutoCorrect list for
Persian (Farsi) language. The old version was created to resolve
i#95322, eleven years ago.

In the old file, there were over 6,000 words in the list, nearly 2/3
of which were unnecessary and essentially did not belong to the
database. It would be better to move them to spell-checking.

Change-Id: Ibe3a2e07ebecdf4a541a5cfacf6fefa8ade4e300
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181004
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>
(cherry picked from commit e142387ffd70d357c0830a9f07d59e4a5b5135f5)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181356
Reviewed-by: Hossein   <hossein@libreoffice.org>
Reviewed-by: افشین ‌ <afshin@libreoffice.org>
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>