tdf#164989: Disallow combining characters when a hex already appeared
We either expect a single (maybe combined) character to convert into "U+"
notation, or a sequence of hexadecimal numbers (possibly with U+) for
conversion to characters. If we already saw a hexadecimal ASCII character,
it can't be preceded by a character combining with it, so stop as soon as
such an unexpected character appears.

Change-Id: Ic480fe8f173240eb263d5a77286b149c933049a6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181007
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit 6d78861299331fdfa51a070ced1ce498ceff342f)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181029
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>