tdf#165002 sidebar a11y: Set a11y names for "Styles" toolbar items in .ui
Similar to how

    Change-Id: I8c30b6b67366d1839acb725bc4af012c276a5ee2
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Feb 3 08:41:49 2025 +0100

        tdf#165002 sidebar a11y: Set a11y names for styles tool buttons

        For the toolbar buttons to switch between the different
        styles categories ("Paragraph Styles", "Character Styles",...)
        in the "Styles" sidebar deck, set the tooltip text also as
        the accessible name.

        This makes the Orca screen reader with announce them
        (e.g. "Paragraph styles, toggle button, pressed") when using
        the gtk3 VCL plugin instead of just saying "toggle button", which
        doesn't give the screen reader any clue what those buttons are for.

did for those toolbar buttons whose tooltips etc. are set in
C++ code, also set a11y names for those that are handled in
the .ui file and set the text used for the tooltip also
as the accessible name, so Orca can announce those as well.
This makes Orca also announce the items in the right-hand
side toolbar of the "Styles" sidebar deck when using the
gtk3 VCL plugin.

Change-Id: Iaf68ec56049feceaf27be5f62cf631a81bf0f50e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181039
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins