wsl-as-helper: make liblangtag more resilient (MSYS path mangling)
different versions of git-bash come with different msys library that behave
slightly differently when it comes to the path mangling that occurs. When msys
encounters something that looks like a path, but the path doesn't exist, then
it prefixes that path using the home of the git-bash installation. That path is
using the long form and since git-bash usually is installed to Program Files
the build then breaks because the path arguments aren't quoted.

The path-mangling can be prevented by starting the path with a double-slash, so
--prefix=//usr/local will take care of those instances where git-bash from
Visual Studio works, but git-bash installed via winget fails.
Luckily that seems to be the only case where the differences actually show
themselves.

Change-Id: I702da05b34dc153267f7251810a1c93babf66022
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181061
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
(cherry picked from commit fc4445dc365bf121715d189b0f769d39660c1411)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181091
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>