ScriptForge (SFDocuments) new MenuHeaders property
The %enuHeaders property returns,
as an array/tuple of strings the list
of the menus present in the menubar.

This is a helper for the CreateMenu()
method:
- check if the menu exists already to
  avoid duplicates
- ease the choice of the Before argument

The property is applicable to next classes:
- Document, and its subclasses
  Base, Calc, Writer, FormDocument
- Datasheet

The property is made available to Basic
and Python user scripts.

This new property will require an update
of the user documentation/help

Change-Id: Ia14daf6c58973e06d576b350e416eb178f70feb3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181065
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins