tdf#165002 sidebar a11y: Set a11y name on sidebar deck buttons
For sidebar deck tool buttons, set  the accessible name
for the actual button, not the toolbar that contains it.

That makes the Orca screen reader announce the name
when the button receives focus instead of just announcing
that a button received focus, leaving it unclear for the
user what this button is about.

Change-Id: I80b3ed45b57ba7cf10b7a46bb0506fdccde55f27
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181073
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>