sidebar: Have sidebar deck tooltip only set on its button
Drop the tooltip from the toolbar that contains the
button.
This is already set for the button as well since

    commit 9fdc7384eff0e9627de2d9c2b59a7716f9e8af25
    Date:   Tue Dec 15 14:33:32 2020 +0000

        tdf#138938 put the tooltip on the toolbar item instead of the toolbar

, so setting it on the toolbar in addition seems unnecessary.
(The tooltip showed fine in a quick test with both, gtk3
and gen as expected.)

    commit bb1d48498c9fdedf8e78344765f6c6eda6f4707c
    Date:   Wed May 18 22:22:06 2022 -0800

        tdf#84502 related: Show keyboard shortcut in sidebar tab bar tab button

had added the shortcut in the tooltip text set for the button,
but not the toolbar, but having these two different variants
doesn't look intentional to me.

Change-Id: I6bdc72d00b33c8a970777ed94a6911042db309ca
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181074
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins