tdf#159029 sw: layout: fix cell vertical orientation after split
The vertical alignment of cells is implemented in SwCellFrame::Format(),
but it is disabled if IsRebuildLastLine() is true (apparently since the
time when splitting table rows was first implemented in commit
90b73c7b667b06d42f5e37196d25a85b6c787f33 "INTEGRATION: CWS cellbreak")
which is in 2 different cases:

1. when another line fits on the page, around a call to
   lcl_RecalcRow(pLastLine) after RemoveFollowFlowLine() - this is the
   relevant case when the paragraph is deleted

This case used to work by accident before commit
068c133ac41c97652909b88c432e3b73010efc3e, when after
RemoveFollowFlowLine() of row 406 the text frame moved to the top, then
a Split of the row after the one containing the last part of the merged
cell with the problem was attempted and failed, and in lcl_RecalcRow()
after the Split the alignment was fixed; this commit changed it so that
the first row frame 6 had an invalid position, which caused an earlier
correction of the position of all the cells and text frames during the
lcl_RecalcRow() after joining the follow, where m_bIsRebuildLastLine was
set.

2. after a successful Split() with bTryToSplit=true, i.e., a row was
   split - this is the relevant case on Undo of the paragraph delete

This actually was a problem already before the above mentioned commit,
but then it could be "fixed" by clicking on the cell, which stopped to
work.

Here there is no RemoveFollowFlowLine() at the start, but then again a
failed attempt to Split a row followed by a successful Split between
rows; before the 2nd Split the text frame is correctly positioned and
valid, but then lcl_RecalcSplitLine() calling
lcl_ShrinkCellsAndAllContent() moves the text frame up.

In both cases, explicitly invalidate the PrtArea of cells in the last
row where a vertical orientation is set.

(regression from commit 068c133ac41c97652909b88c432e3b73010efc3e)

Change-Id: I7d2e85bb76a4b77c7871d759abbebd6abeb31282
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181113
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins
(cherry picked from commit ee482e856143ee62b37869411a64c36a1d78c7cc)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181119
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>