tdf#165035 gtk3 a11y: Set relations for combobox's edit
As described in

    commit 94c97ecdbfa606401fb53685048731128186672b
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Apr 18 18:59:46 2024 +0200

        gtk3 a11y: Set a11y relations for custom combobox impl

, the gtk3 VCL plugin currently uses a custom combobox
implementation.

The above commit implemented that accessible relations set
for the original GtkComboBox in the .ui file
are taken over to the custom combobox implementation's
toggle button.

In case of an editable combobox, also set those relations
for the box containing the combobox's GtkEntry.
(Move the existing logic to set the role up to have everything
in once place.)

This makes the Orca screen reader announce the associated
label's text also when the entry receives focus, not
only when the toggle button does, e.g. in the
"Format" -> "Character" dialog, "Font" tab, for the
"Style" combobox.

Change-Id: I2ece9177355aa18e43e0826bb07b6291944e8912
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181093
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>