treeview a11y: Don't report header twice
AccessibleTabListBox is the a11y class for
SvHeaderTabListBox. That widget only contains
the table without the header, while the (column)
header is implemented by HeaderBar which is a
sibling of the SvHeaderTabListBox, see the
"GtkTreeView" case in VclBuilder::makeObject.

The header bar is already part of the a11y
tree as a11y class for the HeaderBar.

The fact that AccessibleTabListBox::getAccessibleChild
was reporting that as a child as well resulted in the
header bar (an a11y object with Table role) to appear
in the a11y hierarchy seen in Accerciser twice when
using the qt6 VCL plugin. (And clicking on it would
result in an area to be highlighted that is not part
of the area that the parent a11y object occupies.

No longer report it as a child of the AccessibleTabListBox.

The above can be observed with the "Tools" -> "Options"
-> "Advanced" -> "Open Expert Configuration" dialog
and the qt6 VCL plugin on Linux.

Change-Id: If3099e12dd84cd4f3c420e94c377182c87c65106
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181137
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>