treeview a11y: Fix parent-relative coordinates
SvHeaderTabListBox::calcTableRect and
SvHeaderTabListBox::GetFieldRectPixel get used when
parent-relative coordinates are needed in
AccessibleBrowseBoxTable.

Parent-relative coordinates are meant to be
relative to the immediate parent. The immediate
parent of the AccessibleBrowseBoxTable is the
AccessibleTabListBox, so there's no need to
take the relative position of the
AccessibleTabListBox parent within its parent
(i.e. the AccessibleBrowseBoxTable's grand-parent)
into account.

Therefore, drop that extra offset.

Without this commit, the highlighted area was too
far down when clicking through the cells in the
a11y tree of the treeview in the "Tools" -> "Options"
-> "Advanced" -> "Open Expert Configuration" dialog
in Accerciser when using the qt6 VCL plugin on Linux
with an upcoming change to let
AccessibleBrowseBoxBase derive from
OAccessibleComponentHelper which makes use of the
parent-relative coordinates instead of the
screen coordinates currently reported by
AccessibleBrowseBoxBase::getLocationOnScreen.

Change-Id: Ie904701a0a715b22f210361eea3ced969b13c69f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181138
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>