browsebox a11y: Deduplicate 2 cases effectively doing the same
There's no need to distinguish between the case where
`pParent` is set (i.e. parent-relative coords were requested)
and where it is not, as only the height, and not the position
are used, and the size is the same for the two rectangles retrieved
via Window::GetWindowExtentsRelative and
Window::GetWindowExtentsAbsolute, so simply always use the latter.

(The distinction further down is relevant, however,
because it uses the top-left point of the rectangle,
not its size.)

Change-Id: I3362e1493112236e08600f66f7c772d339919e7e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181139
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins