browsebox a11y: Drop incorrect parent-relative pos conversion
The only thing that changes when parent-relative coordinates
(and not screen coordinates) are used is that the origin
of the rectangle is different, which is already handled below.

Passing bRelToBrowser=false to BrowseBox::GetFieldRectPixel
is incorrect and resulted in the position for BrowseBox
headers cells being incorrectly reported on the a11y layer
when parent-relative coordinates are used.

This was observed with an upcoming commit to let
AccessibleBrowseBoxBase derive from OAccessibleComponentHelper
and could be seen e.g. with the table seen in the
Base window that shows up when selecting
"Create Table in Design View..." in the "Tables" tab after
creating a new database. (Highlighted position for
those was wrong in Accerciser when selected in the
a11y tree view there when using the qt6 VCL plugin.)

Change-Id: I4c68748db0f09787037ac1ef023febaf01d1156e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181145
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins