browsebox a11y: Fix parent-relative cell coords
Parent-relative coordinates are meant to be
relative to the immediate parent. The immediate
parent of the cells is the table, and the immediate
parent of the AccessibleBrowseBoxTable
is the AccessibleBrowseBox (which is the a11y
class for BrowseBox).
Therefore:

* Don't add the position of the BrowseBox inside
  of its parent widget when relative coords are
  requested.

* In AccessibleBrowseBoxCell::implGetBoundingBox,
  subtract the relative position of the table
  from the result of BrowseBox::GetFieldRectPixel,
  because BrowseBox::GetFieldRectPixel returns the
  relative position within the BrowseBox (i.e.
  the parent's parent), not within the accessible
  parent.

Without this commit, the wrong area would be highlighted
e.g. for the following scenario when using the gtk3 VCL
plugin:

1) Start base, create a new database
2) With the "Tables" section enabled, click on
   "Create Table in Design View..."#
3) Start Accerciser
4) in Accerciser's treeview of the LO a11y, navigate
   to the a11y object of the (non-header) table cells

Result: Incorrect area is highlighted on screen, not the
one actually covered by the cell.

See also

    commit 589ead3a6ae776fdbcb426f2817fbb94933a8bf7
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Feb 4 17:16:27 2025 +0100

        treeview a11y: Fix parent-relative coordinates

for a similar case.

Change-Id: I45154fcb8365d40984b34ea750cff621ddd0331f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181152
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>