browsebox a11y: Let AccessibleBrowseBox implement XAccessible
Let AccessibleBrowseBox implement the XAccessible
interface and return a reference to self in
its XAccessible::getAccessibleContext implementation.

The AccessibleTabListBox subclass was already doing
this. Drop the override (that didn't hold the mutex
or check whether the accessible is still alive).

This also prepares for merging AccessibleBrowseBoxAccess
into AccessibleBrowseBox into an upcoming commit.

Change-Id: I5cc292e86e2549fa15e00edcc417308d18335b9e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181201
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins