a11y: Move AccessibleBrowseBox* headers to global vcl include dir
AccessibleBrowseBox and related classes are the a11y
implementations for both, SvHeaderTabListBox
(see SvHeaderTabListBox::CreateAccessible) and
BrowseBox (see BrowseBox::CreateAccessible).

The former is in the vcl module, but the latter
is in svtools. Move the headers to include/vcl/accessibility
to be able to use the classes directly from both modules
in upcoming commits instead of having to use the AccessibleFactory
to create the accessible objects.

Change-Id: Idfd5cf2d605f8413f4fa4091ae7c5e4df8cc0cfb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181202
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>