tdf#164745 fix SfxStateCache
Seems I did too much with tdf#162666 by also trying to
remember invalid and disabled states in pLastItem, so
went over it again. Also corrected SetVisibleState to
also take action when disabled item.

Not urgently necessary, but stumbled over it: The items
used for INVALID_POOL_ITEM and DISABLED_POOL_ITEM were
based on the same class. Not a problem technically,
these are 'special' static items that represent a state,
but comparing them would have claimed these to be equal
what should not be the case. Thus I based each of these
on an own derivation of SfxPoolItem.

Change-Id: Ic9dd1a21772bb20b25ed8f7fa929da74edb79e42
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181194
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>