browsebox a11y: Drop AccessibleBrowseBoxAccess
So far, AccessibleBrowseBoxAccess was the XAccessible
implementation for the BrowseBox (in svtools) while
AccessibleBrowseBox is the XAccessibleContext implementation
that AccessibleBrowseBoxAccess::getAccessibleContext() returns.

AccessibleBrowseBoxAccess has some additional methods that
call the corresponding AccessibleBrowseBox methods.

Drop AccessibleBrowseBoxAccess and instead let
AccessibleBrowseBox be the XAccessible and XAccessibleContext
implementation for BrowseBox at the same time.

AccessibleBrowseBox already implements the XAccessible interface
(and AccessibleBrowseBox::getAccessibleContext() returns a reference
to self) since

    Change-Id: I5cc292e86e2549fa15e00edcc417308d18335b9e
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Feb 5 16:45:43 2025 +0100

        browsebox a11y: Let AccessibleBrowseBox implement XAccessible

        Let AccessibleBrowseBox implement the XAccessible
        interface and return a reference to self in
        its XAccessible::getAccessibleContext implementation.

        The AccessibleTabListBox subclass was already doing
        this. Drop the override (that didn't hold the mutex
        or check whether the accessible is still alive).

        This also prepares for merging AccessibleBrowseBoxAccess
        into AccessibleBrowseBox into an upcoming commit.

Now that AccessibleBrowseBox is always its own "creator"
(i.e. the XAccessible whose associated XAccessibleContext it
is), there's no more need to keep track of who created it, so
drop AccessibleBrowseBox::m_aCreator and simply use `this`
instead.

Change-Id: I51099442d600dce0db78a78d37b1f4709c6f95b6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181217
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins