tdf#155050 Replaces the Bibliography entries with Citation
The patch replaces Bibliography entries in Tables Of Contents and Index and bibliography dialog box with Citation.

Left some of the entries in sw/uiconfig/swriter/ui/bibliographyentry.ui and sw/uiconfig/swriter/ui/createauthorentry.ui unchanged as per the discussion on the bug page, where it was mentioned not to modify them.

Change-Id: I34c0051806a799b50d546841960c5d5a04b60196
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181224
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>