tdf#164775 PPTX export: fix connector shape regression
A connector shape can be connected to a shape, if the end of connector shape is released inside the target shape, it will automatically attach to the best possible connection point.
It takes into account the shape type, geometric properties, and
user-defined glue points.
In this case, if we export the ODP file to PPTX, we need to use the GetConnectorRule() method to get the correct connection index(idx).

this regression is
caused by commit 4d153517183193f468dee9148c94fe9d874bacb3
(tdf#149697 PPTX export: fix changing place of connection points).

Change-Id: I8eda059e72f595edc62fd5bc7f3e6466988af5bb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181226
Tested-by: Jenkins
Reviewed-by: Nagy Tibor <tibor.nagy.extern@allotropia.de>