browsebox a11y: Report correct child index for cells
No longer add any offset when a cell's child index
is calculated in
AccessibleBrowseBoxTableCell::getAccessibleIndexInParent.

For SvHeaderTabListBox, that offset was already 0 anyway.

For BrowseBox, that offset was 3 (BBINDEX_FIRSTCONTROL),
presumably so that the first cell comes after
* the column header (BBINDEX_COLUMNHEADERBAR, 0)
* the row header (BBINDEX_ROWHEADERBAR, 1)
* the table (BBINDEX_TABLE, 2)

However, that doesn't make sense, because the row and
column headers and the table are children of the
AccessibleBrowseBox, not the AccessibleBrowseBoxTable,
and the AccessibleBrowseBoxTableCells are children
of the AccessibleBrowseBoxTable, not the AccessibleBrowseBox,
so whatever children the AccessibleBrowseBox has is irrelevant
for the cell's child index.

The incorrect index could be seen like this:

1) Start Base, create a new database
2) With the "Tables" section enabled, click on
   "Create Table in Design View..."
3) Start Accerciser
4) In Accerciser's treeview of the LO a11y, navigate
   to the accessible object for the table (which has
   an accessible name of "Table").
5) Request the child index from the child at index 0

Without this change in place, this is incorrect (0 != 3):

    In [2]: acc.get_child_at_index(0).get_index_in_parent()
    Out[2]: 3

With this change in place, this is now 0 as expected:

    In [3]: acc.get_child_at_index(0).get_index_in_parent()
    Out[3]: 0

This also makes Accerciser's bookmark feature work for
these table cells, i.e. creating a bookmark for a cell
and then jumping to that bookmark later will now select
the correct cell in the treeview again instead of a
different one.

Change-Id: Ic1a074cf604a0d0342dd9e2d4b62b58866b78351
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181237
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>