a11y: Move EditBrowseBoxTableCell to svtools
While the AccessibleBrowseBoxCell base class
and other AccessibleBrowseBox* classes are
used for both, SvHeaderTabListBox (in vcl)
and BrowseBox (in svtools), EditBrowseBoxTableCell
is only used for the latter, in order to make
the currently active edit control accessible.

Therefore, move the class to the svtools library,
where BrowseBox is located as well.

This is also in preparation of reworking
how accessibility for BrowseBox's currently active
control is implemented.

Change-Id: Ia290df4b497baa66f98c1065c2425a0e2e079d13
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181267
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>