cool#11103 sc: respect document specific settings
The purpose of this CheckConfigOptions function is to reset
formula separators, when there is a clash with decimal separator.
However, ignoring other settings that are coming from the document
can cause issues, like that we saw with LOKit, where
StringRefAddressSyntax of a German XLSX document was reset to default
CONV_UNSPECIFIED, and function =INDIREKT("Sheet1!A1") which was
according to CONV_XL_A1 stopped working, and gave error #BEZUG (#REF).

Change-Id: I7693deed436ad4e54569a21cc3faeaafceeb459c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181250
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181287
Tested-by: Jenkins