tdf#164712: ITEM regression: Fix SfxItemSet::Differentiate
For SfxItemSet::Differentiate and ::Intersect we had the strange
case that dependent of the former possible same WhichRanges two
different things were done, see the comments I had added to those
methods when adapting to ITEM changes.
This error now has given a hint that one of those was the intended
one, the one that ALSO does that for item in state
SfxItemState::DISABLED or SfxItemState::INVALID.
Thus I now adapted these methods to (hopefully now) do the right
thing in all cases.

Change-Id: I276da418d663d37942530563c4bfaf0a46a93e26
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181258
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>