tdf#165062 qt: Make only floating window/popup widget bg translucent
For QtWidget, only set the window attributes needed for
translucent background for floating windows and popups,
not all QtWidgets.

This fixes rendering issues when resizing a Window
with Notebookbar (Groupedbar Compact) active (tdf#165062).

Setting those flags was introduced in

    commit 9ea767cb568cef1b142190d2adb0e301baa382e2
    Date:   Thu May 26 13:48:38 2022 +0200

        tdf#132350 Qt implement SalMenu button interface

whose commit message explains:

>  This also includes the transparency flags for our QtWidget, so the
>  updater "bubble window" is properly alpha-masked / shaped.

Indeed, the BubbleWindow doesn't properly use alpha
on X11 if those flags are not set  in a quick test with demo change [1] applied
locally to trigger the update notification.

(Even without those flags set, the window still uses
a translucent background on Wayland in my tests. It has
a different problem of being misplaced there, but that's
a completely independent issue, see tdf#158994.)

[1] https://gerrit.libreoffice.org/c/core/+/106922/4

Change-Id: Ibda381985c0dc562d7548656146f99f01fca1964
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181292
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>