hopefully fix the PDF export of alpha once and for all
The root of the problem, is that at the bottom of the call-chain,
we want to export transparency. We were trying to patch this in the
middle of call-chain, by inverting. But that is unreliable because
sometimes we don't always traverse the middle of the call-chain
once.

So remove all of the various inverting, and just invert at the
bottom, when we want to output transparency data.

Tested this with the various bugs linked in the comments I removed.

Change-Id: I880ab3a3e6c44f8531893f7405a33a331e4b4295
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181316
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins