vcl: Replace use of MYTABMASK in SvTabListBox::SetTabAlignCenter
Seeing

    constexpr SvLBoxTabFlags MYTABMASK =
        SvLBoxTabFlags::ADJUST_RIGHT | SvLBoxTabFlags::ADJUST_LEFT | SvLBoxTabFlags::ADJUST_CENTER | SvLBoxTabFlags::FORCE;

, it is not particularly obvious what the idea behind that
set of flags is.

Introduce SvLBoxTabFlags::ADJUST_FLAGS for the adjustment
flags and use that one in SvTabListBox::SetTabAlignCenter instead,
where the purpose is to clear existing alignment flags before setting
SvLBoxTabFlags::ADJUST_CENTER.

Explicitly clearing SvLBoxTabFlags::FORCE (as happened
before) isn't necessary because that flag is explicitly
set afterwards (so it doesn't make any difference whether
or not it is cleared first).

Change-Id: Ib26e103185c4a79cae97c8ef38b99a02c66c6368
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181339
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>